Implemented a FireAndForget service for asynchronous database writes.

+++ src/Core/MessageSender.Domain/Contracts/ISmsRepository.cs
@@ -5,5 +5,5 @@ namespace MessageSender.Domain.Contracts;
 public interface ISmsRepository
 {
     Task<long> InsertSmsAsync(Guid clientId, string phoneNumber, CancellationToken cancellationToken = default);
-    Task UpdateSmsContentAsync(int smsId, string content, CancellationToken cancellationToken = default);
+    Task UpdateSmsContentAsync(long smsId, string content, CancellationToken cancellationToken = default);
 }
